backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/RAG Dependencies
langchain==0.1.4
langchain-community==0.0.16
sentence-transformers==2.3.1
faiss-cpu==1.7.4
pypdf==3.17.4
tiktoken==0.5.2

# Utilities
aiofiles==23.2.1
websockets==12.0
pytesseract==0.3.10
pdf2image==1.17.0
